Created seperate file for Errors
Transferred over errors
